# backend/api/src/main/java/top/zhaizz/animetracker/user/controller/AdminUserController.java
package top.zhaizz.animetracker.user.controller;

import jakarta.validation.Valid;
import lombok.RequiredArgsConstructor;
import org.springframework.web.bind.annotation.*;
import top.zhaizz.animetracker.common.ApiResponse;
import top.zhaizz.animetracker.common.PageResult;
import top.zhaizz.animetracker.user.dto.UpdateRoleRequest;
import top.zhaizz.animetracker.user.service.UserService;
import top.zhaizz.animetracker.user.vo.UserVO;

/**
 * 管理员用户管理控制器
 */
@RestController
@RequestMapping("/api/admin/users")
@RequiredArgsConstructor
public class AdminUserController {

    private final UserService userService;

    /**
     * 分页查看所有注册用户（不返回密码字段）
     */
    @GetMapping
    public ApiResponse<PageResult<UserVO>> listUsers(
            @RequestParam(defaultValue = "1") int page,
            @RequestParam(defaultValue = "20") int size) {
        return ApiResponse.success(userService.listUsers(page, size));
    }

    /**
     * 修改指定用户的角色
     * <p>
     * `@Valid` 定义校验规则,在 UpdateRoleRequest 类中使用 @NotBlank、@NotNull、@Size 等注解定义校验规则
     */
    @PutMapping("/{id}/role")
    public ApiResponse<UserVO> updateUserRole(
            @PathVariable Long id,
            @Valid @RequestBody UpdateRoleRequest request) {
        return ApiResponse.success(userService.updateUserRole(id, request.getRole()));
    }
}